pyproject.toml
[project]
name = "traffic-simulator"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "David Lin", email = "davidzylin@outlook.com"},
]
dependencies = ["numpy>=2.2.2", "click>=8.1.8", "matplotlib>=3.10.0", "pyyaml>=6.0.2", "pydantic>=2.10.6"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
traffic-simulator = "traffic_simulator.runner:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[dependency-groups]
lint = [
    "ruff>=0.9.4",
]
